Fix PNaCl configure build
Moving all Emscripten passes to JSBackend was a bad idea because it caused a layering violation and made libLLVMNaClTransforms.a fail to compile. Move them back to lib/Transforms/NaCl/ as in the Emscripten repo, and also move ExpandI64 which should have been there in the first place (and fix the corresponding compile errors).

R=dschuff@chromium.org, azakai@mozilla.com, sunfish@mozilla.com
BUG= https://code.google.com/p/nativeclient/issues/detail?id=4102
TEST= make

Review URL: https://codereview.chromium.org/1051033004