Silence ccache-induced warnings about unused arguments.
This applies only to the cmake build.

In some people's build environments (but unfortunately not mine), setting the CCACHE_CPP2 envvar is not sufficient to get rid of the spurious "unused argument" warnings.  This turns into an actual problem when we compile with -Werror.

BUG= none
R=jfb@chromium.org

Review URL: https://codereview.chromium.org/1057023002