Remove raw_ostream localmod
This was added by @juj in e8fdaf5a2f09e335bdc5a6689b298e54337cc737 to fix the VS2010 build. It's either not needed anymore because the diff from LLVM upstream on this file is empty, or it's instead needed in upstream LLVM (I doubt it, since LLVM does build correcting on MSVC).