Use AEABI runtime support functions by default for ARM NaCl
ARM EABI specifies a set of functions for the runtime support library, with
different names than those used by other platforms, as well as a few functions
not present elsewhere. Use of these has been disabled in PNaCl and nacl-clang,
but it's better to be more conventional, especially for nacl-clang.

We have a local flag that disables the special ARM flavor (which is given to
llc by pnacl-translate) but there's also an explicit check for NaCl which
changes the behavior for nacl-clang. Remove the explicit check to change
nacl-clang, and keep the flag for PNaCl for now.

R=jvoung@chromium.org
BUG= https://code.google.com/p/nativeclient/issues/detail?id=4140

Review URL: https://codereview.chromium.org/1076003002