Add some configure/CMake options for enabling/disabling Subzero targets.
Mirrors the LLVM_TARGETS_TO_BUILD, but kept separate since
(a) the spellings are different
(b) we may keep X8632 separate from X8664 unlike LLVM

This setups up an xmacro file to enumerate the enabled targets.

NaCl change:
https://codereview.chromium.org/1082953002/

Subzero change:
https://codereview.chromium.org/1075363002/

This change has to be atomic w/ the NaCl change since the
default is "none" which blows up.

An alternative may be to move the header to Subzero, and
"configure" stuff via -D in the CXXFLAGS, filling in the
.def file that way if possible.

BUG= https://code.google.com/p/nativeclient/issues/detail?id=4076
R=stichnot@chromium.org

Review URL: https://codereview.chromium.org/1084753002