Add fast isel pseudo for materializing 32 bit immediates with movt
The MOVi32imm ARM pseudo instruction expands to a movw+movt pair to
materialize a 32 bit immediate. However it cannot be selected by fast
isel because it has a pattern as an operand instead of a simple type
(unlike its thumb2 counterpart).

ARMFastISel::ARMMaterializeInt will normally just fall back to constant pool
load for this, but these are disabled for NaCl, so
ARMFastISel::ARMSimplifyAddress fails to materialize a constant for selecting
loads and stores (and does not fall back to dag isel).

So, introduce another pseudo instruction which is simple enough pattern such
that tablegen will generate a fast isel selection pattern for it. Then
ARMMaterializeInt will not fall back to a constpool load.

R=jfb@chromium.org, jvoung@chromium.org
BUG= https://code.google.com/p/nativeclient/issues/detail?id=4114

Review URL: https://codereview.chromium.org/1097393009