ARM MC: Create Immediate operands using 12-bit encoded form
r225266 changed MC such that immedates are store in encoded form rather than
full unsigned form, so we encode the bic masks using the getSOImmVal utility
function when creating them.
The disassembler also changed to print the immediate values as signed rather
than unsigned, so the tests are updated accordingly (after manually verifying
that the instruction encoding is what we want).