Add (unsupported experimental) feature allowing byte aligned bitcode.
This is the first step of allowing PNaCl bitcode files to generate and
accept byte aligned bitcode records. To write bitcode files with byte
alignment, one uses the command-line argument
'--align-bitcode-records'. When such bitcode files are written, a
corresponding flag is added to the bitcode header so that the reader
knows whether it should align bitcode records. This allows old PNaCl
bitcode files to continue to work.

BUG=https://code.google.com/p/nativeclient/issues/detail?id=4164
R=jvoung@chromium.org

Review URL: https://codereview.chromium.org/1122423005