Extend the munging bitcode records to generate text records.
This CL modifies the writer of munged bitcode to be able to generate
simple textual representations of records. The goal is to generate a
simple form for which tokenizing fuzzing can be applied.

The simplified form removes all abbreviations, assuming that the
default abbreviations should be used to encode records. A record is a
sequence of integers, separated by commas, and terminated with a
semicolin. For readability, a newline is added after the terminator so
that each line corresponds to a record.

Also modifies pnacl-bcfuzz to be able to convert PNaCl bitcode files
into thier textual form.

BUG=https://code.google.com/p/nativeclient/issues/detail?id=4169
R=jvoung@chromium.org

Review URL: https://codereview.chromium.org/1179773003