Modify pnacl-{llc,thaw} to read textual bitcode records.
This CL adds a reader of textual bitcode records. The goal is to
allow a simple form that can be mutated via fuzzing.

This CL is the reader for the textual output generated by
CL https://codereview.chromium.org/1179773003 .

The modifications to pnacl-llc were made in such a way as to
not introduce this concept to the browser translator.

BUG= https://code.google.com/p/nativeclient/issues/detail?id=4169
R=jvoung@chromium.org

Review URL: https://codereview.chromium.org/1191493006