Cherrypick upstream r240877 This fixes the naclports binutils build.
[LoopVectorize] Pointer indicies may be wider than the pointer

If we are dealing with a pointer induction variable, isInductionPHI
gives back a step value of Stride / size of pointer.  However, we might
be indexing with a legal type wider than the pointer width.
Handle this by inserting casts where appropriate instead of crashing.

This fixes PR23954.

git-svn-id: https://llvm.org/svn/llvm-project/llvm/trunk@240877 91177308-0d34-0410-b5e6-96231b3b80d8

R=jfb@chromium.org
BUG= https://code.google.com/p/nativeclient/issues/detail?id=4202

Review URL: https://codereview.chromium.org/1215973004.