Fix code to handle bad blockinfo block abbreviation index bits.
Fixes code that handles a bad specification of the number of bits
to used in a blockinfo block, when writing out the corresponding
binary form of a list of bitcode records.

It should be noted that this doesn't effect the production
translator. It generates sufficient bits to handle the
problem. Rather, it fixes test cases (where it is easy to
make a mistake).

BUG= https://code.google.com/p/nativeclient/issues/detail?id=4215
R=jvoung@chromium.org

Review URL: https://codereview.chromium.org/1219543002.