SimplifyAllocas: Handle any noop cast (not just bitcast) for dbg.declare
Look through any CastInst (to handle e.g. inttoptr/ptrtoint) to find the
alloca for a dbg.declare intrinsic. Also, if an alloca cannot be found,
do not modify the dbg.declare (the backend may or may not handle it, but
if not it will be silently dropped).

R=jvoung@chromium.org
BUG= https://code.google.com/p/nativeclient/issues/detail?id=4203

Review URL: https://codereview.chromium.org/1216073010.