retire RELOOPERDLL_API to fix MinGW shared build
With __declspec(dllexport), only the relooper APIs will be exported in
the final DLL when building with MinGW. As a result, the shared build
(-DBUILD_SHARED_LIBS=ON) breaks because LLVMJSBackendCodeGen.dll is not
exporting all necessary symbols. The patch simply removes this
attribute. MSVC static build is supposed to be unaffected because the
relooper APIs now go in LLVMJSBackendCodeGen.lib the same way as other
LLVM symbols, and MSVC shared build is just not supported by LLVM.