fix 64-bit wasm atomic compareExchange (#244)
This cannot be just lowered to a wasm instruction as the wasm version receives one pointer, while the LLVM IR version we lower receives two pointers, and may modify both. This is basically the same difference between

* The c++ API, which matches the libatomic call clang emits for asm.js/asm2wasm: https://en.cppreference.com/w/cpp/atomic/atomic/compare_exchange
* The wasm API: https://github.com/WebAssembly/threads/blob/master/proposals/threads/Overview.md#example
We need to implement the high-level one from LLVM IR using the lower-level one. for now, just use the asm.js version to work as well as that one. This fixes build breakage in asm2wasm, and likely fixes incorrect runtime in asm.js.

PR also contains a necessary for for a bitcast on a call, which casts away an i64 - we didn't handle that, but clang emits it for that c++11 atomic.